refactor: fix norm
