+++ Makefile
@@ -44,6 +44,7 @@ SRC			= $(SRC_DIR)/main.c \
 				$(SRC_DIR)/utils/error.c \
 				$(SRC_DIR)/vec3/arithmetic.c \
 				$(SRC_DIR)/vec3/geometric.c \
+				$(SRC_DIR)/vec3/metric.c \
 				$(SRC_DIR)/viewport/init.c
 
 OBJ			= $(patsubst $(SRC_DIR)/%.c, $(BUILD_DIR)/%.o, $(SRC))
