+++ include/vector.h
@@ -6,7 +6,7 @@
 /*   By: reasuke <reasuke@student.42tokyo.jp>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/04/27 11:00:30 by yliu              #+#    #+#             */
-/*   Updated: 2025/05/10 23:35:05 by reasuke          ###   ########.fr       */
+/*   Updated: 2025/05/10 23:39:50 by reasuke          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -45,9 +45,11 @@ t_vec3	vec3_clamp(t_vec3 a, double min, double max);
 // geometric.c
 double	vec3_dot(t_vec3 a, t_vec3 b);
 t_vec3	vec3_cross(t_vec3 a, t_vec3 b);
-double	vec3_length(t_vec3 a);
 t_vec3	vec3_normalize(t_vec3 a);
-bool	vec3_is_parallel(t_vec3 a, t_vec3 b);
 t_vec3	vec3_reflect(t_vec3 incident, t_vec3 normal);
+
+// metric.c
+bool	vec3_is_parallel(t_vec3 a, t_vec3 b);
+double	vec3_length(t_vec3 a);
 
 #endif
